# Use prctl to enable/ disable perf for lower overhead