--- a/test/runtests.jl
+++ b/test/runtests.jl
@@ -1,7 +1,7 @@
 using LinuxPerf
 using Test
 
-using LinuxPerf: make_bench, enable!, disable!, reset!, reasonable_defaults, counters, EventType, EventTypeExt, parse_groups
+using LinuxPerf: make_bench, enable!, disable!, reset!, reasonable_defaults, counters, EventType, EventTypeExt, parse_groups, Counter, ThreadStats, Stats, enable_all!, disable_all!, scaledcount, isenabled, isrun
 
 @testset "LinuxPerf" begin
 
@@ -21,7 +21,39 @@ using LinuxPerf: make_bench, enable!, disable!, reset!, reasonable_defaults, cou
         end
         g(zeros(10000))
 
-        counters(bench)
+        results = counters(bench)
+        close(bench)
+
+        @test all(x->isenabled(x) && isrun(x), results.counters)
+        @test scaledcount(results.counters[findfirst(x->x.event == EventType(:hw, :cycles), results.counters)]) > 1000
+        @test scaledcount(results.counters[findfirst(x->x.event == EventType(:hw, :branches), results.counters)]) > 250
+        @test scaledcount(results.counters[findfirst(x->x.event == EventType(:hw, :instructions), results.counters)]) > 1000
+
+        true  # Succeeded without any exceptions...
+    end
+
+    @test begin
+        bench = make_bench(reasonable_defaults);
+        @noinline function g(a)
+            enable_all!()
+            c = 0
+            for x in a
+                if x > 0
+                    c += 1
+                end
+            end
+            disable_all!()
+            c
+        end
+        g(zeros(10000))
+
+        results = counters(bench)
+        close(bench)
+
+        @test all(x->isenabled(x) && isrun(x), results.counters)
+        @test scaledcount(results.counters[findfirst(x->x.event == EventType(:hw, :cycles), results.counters)]) > 1000
+        @test scaledcount(results.counters[findfirst(x->x.event == EventType(:hw, :branches), results.counters)]) > 250
+        @test scaledcount(results.counters[findfirst(x->x.event == EventType(:hw, :instructions), results.counters)]) > 1000
 
         true  # Succeeded without any exceptions...
     end