@Zentrik 2024-04-29
I forgot to check if prctl returns an error code, I'll add that in a bit.